Start insert

+++ Factory/Factory/MainWindow.xaml.cs
@@ -105,7 +105,7 @@ namespace Factory
         //INSERT into Firestore method
         private async void InsertFS()
         {
-            //Image path
+            /*//Image path
             string solutionDirectory = Directory.GetParent(baseDir).Parent.Parent.Parent.Parent.Parent.FullName;
             //string filePath = Path.Combine(solutionDirectory, "images/car.jpg");
             string filePath = Path.Combine(solutionDirectory, "images/car.jpg");
@@ -117,28 +117,71 @@ namespace Factory
                 await _storage.UploadObjectAsync(this._bucketName, image, null, fileStream);
 
             }
-            string filePathUrl = Path.Combine(solutionDirectory, "images");
-            //We need the public URL of the uploaded file to tie it to the object in the collection
-            //THIS LINK IS TRANSLATED FROM GS TO ACCESS IT VIA THE CLOUD ! ADD ?alt=media at the END TO CHANGE THE ENCODING TO IMAGE !!!!
-            var url = $"https://firebasestorage.googleapis.com/v0/b/{this._bucketName}/o/{Uri.EscapeDataString(filePathUrl)}%2F{Uri.EscapeDataString("car.jpg")}?alt=media";
-            //https://firebasestorage.googleapis.com/v0/b/{this._bucketname}/o/{image}
+            string filePathUrl = Path.Combine(solutionDirectory, "images");*/
 
-            //string connectionString = "Server=DESKTOP-P1UFSEM;Database=test;Integrated Security=true;TrustServerCertificate=True";
-            Show show = new Show { Name = "ShowN", Description = "ShowD", ImageUrl = url, Category = "Action" };
-            Anime anime = new Anime { Name = "AnimeN", Description = "AnimeD", ImageUrl = url, Category = "Comedy" };
-            Movie movie = new Movie { Name = "MovieN", Description = "MovieD", ImageUrl = url, Category = "Drama" };
+            var picker = new FileOpenPicker();
+            picker.ViewMode = PickerViewMode.Thumbnail;
+            picker.SuggestedStartLocation = PickerLocationId.PicturesLibrary;
+            picker.FileTypeFilter.Add(".jpg");
+            picker.FileTypeFilter.Add(".jpeg");
+            picker.FileTypeFilter.Add(".png");
 
-            CollectionReference watchableNode = _db.Collection("watchables");
-            DocumentReference showRef = watchableNode.Document(show.Name);
-            DocumentReference movieRef = watchableNode.Document(movie.Name);
-            DocumentReference animeRef = watchableNode.Document(anime.Name);
+            Window currentWindow = Window.Current ?? new Window();
+            nint windowHandle = WindowNative.GetWindowHandle(currentWindow);
+            InitializeWithWindow.Initialize(picker, windowHandle);
 
-            await showRef.SetAsync(new Converter<Show>().ToFirestore(show));
-            await movieRef.SetAsync(new Converter<Movie>().ToFirestore(movie));
-            await animeRef.SetAsync(new Converter<Anime>().ToFirestore(anime));
+            var file = await picker.PickSingleFileAsync();
 
-            loadB.Content = "Sent data";
-            File.Delete(filePath);
+            if (file != null)
+            {
+                string imgName = Path.GetFileName(file.Path);
+
+                // Copy the selected file to the images folder
+                string imagesFolderPath = Path.Combine(Directory.GetParent(baseDir).Parent.Parent.Parent.Parent.Parent.FullName, "images");
+                StorageFolder imagesFolder = await StorageFolder.GetFolderFromPathAsync(imagesFolderPath);
+                await file.CopyAsync(imagesFolder, imgName, NameCollisionOption.ReplaceExisting);
+
+                string solutionDirectory = Directory.GetParent(baseDir).Parent.Parent.Parent.Parent.Parent.FullName;
+                string filePath = Path.Combine(solutionDirectory, "images/", imgName);
+                var image = filePath;
+
+                using (var fileStream = File.OpenRead(filePath)) //Otherwise we can't delete the file at the end cause it's still in use by the thread, so it closes the FS
+                {
+                    await _storage.UploadObjectAsync(this._bucketName, image, null, fileStream);
+
+                }
+                string filePathUrl = Path.Combine(solutionDirectory, "images");
+
+                if (!string.IsNullOrEmpty(imgName))
+                {
+                    /*
+                    //We need the public URL of the uploaded file to tie it to the object in the collection
+                    //THIS LINK IS TRANSLATED FROM GS TO ACCESS IT VIA THE CLOUD ! ADD ?alt=media at the END TO CHANGE THE ENCODING TO IMAGE !!!!
+                    var url = $"https://firebasestorage.googleapis.com/v0/b/{this._bucketName}/o/{Uri.EscapeDataString(filePathUrl)}%2F{Uri.EscapeDataString("car.jpg")}?alt=media";
+                    //https://firebasestorage.googleapis.com/v0/b/{this._bucketname}/o/{image}
+                    */
+
+                    //string connectionString = "Server=DESKTOP-P1UFSEM;Database=test;Integrated Security=true;TrustServerCertificate=True";
+
+                    var url = $"https://firebasestorage.googleapis.com/v0/b/{this._bucketName}/o/{Uri.EscapeDataString(filePathUrl)}%2F{Uri.EscapeDataString(imgName)}?alt=media";
+
+                    Show show = new Show { Name = "ShowN", Description = "ShowD", ImageUrl = url, Category = "Action" };
+                    Anime anime = new Anime { Name = "AnimeN", Description = "AnimeD", ImageUrl = url, Category = "Comedy" };
+                    Movie movie = new Movie { Name = "MovieN", Description = "MovieD", ImageUrl = url, Category = "Drama" };
+
+                    CollectionReference watchableNode = _db.Collection("watchables");
+                    DocumentReference showRef = watchableNode.Document(show.Name);
+                    DocumentReference movieRef = watchableNode.Document(movie.Name);
+                    DocumentReference animeRef = watchableNode.Document(anime.Name);
+
+                    await showRef.SetAsync(new Converter<Show>().ToFirestore(show));
+                    await movieRef.SetAsync(new Converter<Movie>().ToFirestore(movie));
+                    await animeRef.SetAsync(new Converter<Anime>().ToFirestore(anime));
+
+                    loadB.Content = "Sent data";
+                    File.Delete(filePath);
+                }
+            }
         }
 
         //Retrieve data from the database (Firestore) method (attached to no button, connected to the ListView)
